+++ packages/blocks/core/CHANGELOG.md
@@ -88,6 +88,7 @@
 - `Comment Author Name` block: support improved.
 - `Comment Date` block: support improved.
 - `Comment Edit Link` block: support improved.
+- `Comments Pagination` block: support improved.
 
 
 ### Automated Tests
@@ -113,6 +114,7 @@
 - Updated E2E tests to check support of `Comment Author Name` block.
 - Updated E2E tests to check support of `Comment Date` block.
 - Updated E2E tests to check support of `Comment Edit Link` block.
+- Updated E2E tests to check support of `Comments Pagination` block.
 
 
 ## 1.3.0 (2025-04-27)